apps/video-processor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
boto3==1.34.0
redis==5.0.0
ffmpeg-python==0.2.0
openai-whisper==20231117
python-dotenv==1.0.0
